requirements.txt
fastapi>=0.116.1
moviepy==1.0.3
openai>=1.98.0
pydantic>=2.11.7
pytest>=8.4.1
python-multipart>=0.0.20
requests>=2.32.4
sqlalchemy>=2.0.42
uvicorn>=0.35.0
oss2>=2.19.1
edge-tts>=7.2.0
python-dotenv>=1.1.1
aiofiles
psutil